Inline _geom.py into _atom.py and _combine.py into _factor.py
Both were single-consumer modules: _geom's copy_cell/shrink_axis were used
only by _atom.py (plus _factor.py, which already depends on _atom.py) and
_combine's emit_combine/combine_tail only by _factor.py. extent_expr becomes
Axis.extent_expr(), a method on the type it reads instead of a free function
needing a shared home.
